worker/requirements.txt
celery
redis
python-dotenv
sendgrid
